Include a copy of the license in repository
According to your `package.json`, you are releasing this under the ISC license. Including a copy of the license in your repository lets GitHub tag the repo with the license. Furthermore, since the ISC license requires people who want to use your code include the copyright notice, including a copy of the license with the correct copyright notice makes it easier for others to comply with the terms of the license.